+++ src/app/controllers/eventos.js
@@ -27,6 +27,19 @@ class Evento {
     })
   }
 
+  update(req, res) {
+    let title = req.params.title.replace(/%20/g, " ")
+    let body = req.body
+        
+    EventoSchema.updateOne({ titulo: title }, { $set: body }, (err, data) => {
+      if (err) {
+        res.status(500).json({ message: 'Houve um erro ao processar sua requisição', error: err })
+      } else {
+        res.status(201).json({ message: 'Evento atualizado com sucesso', data: data })
+      }
+    })
+  }
+
   getWithParams(req, res) {
     let limit = parseInt(req.query.limit)
     let query = {}