+++ src/app/routes/private/routes/eventos.private.js
@@ -5,5 +5,6 @@ const EventoController = require('./../../../controllers/eventos')
 route.post('/criar', EventoController.create)
 route.get('/listar-todos', EventoController.getWithParams)
 route.get('/listar-um/:title', EventoController.getByTitle)
+route.put('/atualizar/:title', EventoController.update)
 
 module.exports =  route