+++ src/app/routes/private/routes/integrantes.private.js
@@ -7,5 +7,6 @@ route.post('/criar', IntegrantesController.create)
 route.get('/listar-todos', IntegrantesController.getWithParams)
 route.get('/listar-um/:nome', IntegrantesController.getByName);
 route.put('/atualizar/:nome',IntegrantesController.update);
+route.delete('/apagar/:id', IntegrantesController.delete);
 
 module.exports = route