+++ CHANGELOG.md
@@ -38,3 +38,4 @@ Log de alterações da aplicação
 * **Integrantes:** Alterando a model de integrante para aceitar email.
 * **Integrantes:** Adição da rota e metodo para editar integrante.
 * **Eventos:** Implementando o método de update e disponibilizando a rota atualizar.
+* **Integrantes:** Adição da rota e metodo para excluir integrante.