+++ CHANGELOG.md
@@ -42,3 +42,5 @@ Log de alterações da aplicação
 * **Sobre:** Implementando o método de update e disponibilizando a rota atualizar.
 * **Sobre:** Implementando metodo de ordenação do Sobre.
 * **Autores:** Implementado o modelo de dados, controller com os métodos getWithParams e create, rotas listar e criar e validador de nome único com a rota do validador.
+* **Publicações:** Implementado o método update e getPublicacoesByTitleWithoutFiles no controller e disponibilizado as rotas getdata e atualizar e modificado o campo autores.
+* **Autores:** Alterado o método getWithParams.